apps/control-plane/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-control-plane"
version = "0.1.0"
description = "Sentinel MCP control plane (personal R&D prototype)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "Jason Lovell" }]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy>=2.0.35",
    "alembic>=1.13.3",
    "psycopg[binary,pool]>=3.2.1",
    "pydantic-settings>=2.4.0",
    "httpx>=0.27.2",
    "redis>=5.0.8",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.2.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
    "pytest-cov",
]

[project.urls]
homepage = "https://github.com/jasonlovell/sentinel-mcp"

[tool.setuptools.packages.find]
where = ["src"]
